requirements.txt
uvicorn~=0.20.0
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124
transformers~=4.44.1
streamlit
nltk
python-telegram-bot~=20.3
requests
azure-ai-formrecognizer==3.2.0
azure-core==1.28.0